backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
pandas>=2.0
openpyxl>=3.0
python-dotenv>=1.0
openai>=0.27.0
